+++ elmyra.ip.access.epo/CHANGES.rst
@@ -9,6 +9,7 @@ development
 - Improve OPS drawing inquiry re. US amendments/corrections
 - Improve"FulltextPRO "session management for error cases
 - Improve placeholders re. gracefulness to WO anomalies like WO2003049775A2 vs. WO03049775A2
+- Improve drawings carousel by using drawings from family members for references (Aktenzeichen) like DE112013003369A5
 
 0.113.0
 =======